transferred /start_recording route over to server.py
