+++ tests/test_main.py
@@ -297,131 +297,6 @@ def test_main__does_not_call_boot_up_function_upon_launch_if_command_line_arg_pa
     spied_boot_up.assert_not_called()
 
 
-@pytest.mark.parametrize(
-    "test_barcode,expected_error_message,test_description",
-    [
-        (
-            "MA1234567890",
-            "Barcode exceeds max length",
-            "returns error message when barcode is too long",
-        ),
-        (
-            "MA1234567",
-            "Barcode does not reach min length",
-            "returns error message when barcode is too short",
-        ),
-        (
-            "MA21044-001",
-            "Barcode contains invalid character: '-'",
-            "returns error message when '-' is present",
-        ),
-        (
-            "M$210440001",
-            "Barcode contains invalid character: '$'",
-            "returns error message when '$' is present",
-        ),
-        (
-            "MZ20044001",
-            "Barcode contains invalid header: 'MZ'",
-            "returns error message when barcode header is invalid",
-        ),
-        (
-            "MA210440001",
-            "Barcode contains invalid year: '21'",
-            "returns error message when year is invalid",
-        ),
-        (
-            "MA200000001",
-            "Barcode contains invalid Julian date: '000'",
-            "returns error message when julian date is too low",
-        ),
-        (
-            "MA20367001",
-            "Barcode contains invalid Julian date: '367'",
-            "returns error message when julian date is too big",
-        ),
-        (
-            "MA2004400BA",
-            "Barcode contains nom-numeric string after Julian date: '00BA'",
-            "returns error message when barcode ending is non-numeric",
-        ),
-        (
-            "MA2004400A",
-            "Barcode contains nom-numeric string after Julian date: '00A'",
-            "returns error message when barcode ending is non-numeric",
-        ),
-    ],
-)
-def test_start_recording__returns_error_code_and_message_if_barcode_is_invalid(
-    test_client, test_barcode, expected_error_message, test_description
-):
-    response = test_client.get(f"/start_recording?barcode={test_barcode}")
-    assert response.status_code == 400
-    assert response.status.endswith(expected_error_message) is True
-
-
-def test_start_recording__returns_error_code_and_message_if_barcode_is_not_given(
-    test_client,
-):
-    response = test_client.get("/start_recording")
-    assert response.status_code == 400
-    assert response.status.endswith("Request missing 'barcode' parameter") is True
-
-
-def test_start_recording__returns_error_code_and_message_if_customer_account_id_not_set(
-    test_client, patched_start_recording_shared_dict
-):
-    patched_start_recording_shared_dict["config_settings"]["Customer Account ID"] = ""
-    response = test_client.get("/start_recording?barcode=MA200440001")
-    assert response.status_code == 406
-    assert response.status.endswith("Customer Account ID has not yet been set") is True
-
-
-def test_start_recording__returns_error_code_and_message_if_user_account_id_not_set(
-    test_client, patched_start_recording_shared_dict
-):
-    patched_start_recording_shared_dict["config_settings"]["User Account ID"] = ""
-    response = test_client.get("/start_recording?barcode=MA200440001")
-    assert response.status_code == 406
-    assert response.status.endswith("User Account ID has not yet been set") is True
-
-
-def test_start_recording__returns_error_code_and_message_if_called_with_is_hardware_test_mode_false_when_previously_true(
-    test_client, patched_start_recording_shared_dict, test_process_manager
-):
-    test_process_manager.create_processes()
-
-    response = test_client.get(
-        "/start_recording?barcode=MA200440001&is_hardware_test_recording=True"
-    )
-    assert response.status_code == 200
-    response = test_client.get(
-        "/start_recording?barcode=MA200440001&is_hardware_test_recording=False"
-    )
-    assert response.status_code == 403
-    assert (
-        response.status.endswith(
-            "Cannot make standard recordings after previously making hardware test recordings. Server and board must both be restarted before making any more standard recordings"
-        )
-        is True
-    )
-
-
-def test_start_recording__returns_no_error_message_with_multiple_hardware_test_recordings(
-    test_client, patched_start_recording_shared_dict, test_process_manager
-):
-    test_process_manager.create_processes()
-
-    response = test_client.get(
-        "/start_recording?barcode=MA200440001&is_hardware_test_recording=True"
-    )
-    assert response.status_code == 200
-    response = test_client.get(
-        "/start_recording?barcode=MA200440001&is_hardware_test_recording=True"
-    )
-    assert response.status_code == 200
-
-
 def test_main__logs_command_line_arguments(mocker):
     expected_command_line_args = ["--debug-test-post-build", "--log-level-debug"]
     spied_info_logger = mocker.spy(main.logger, "info")
@@ -727,14 +602,3 @@ def test_route_with_no_url_rule__returns_error_message__and_logs_reponse_to_requ
     mocked_logger.assert_called_once_with(
         f"Response to HTTP Request in next log entry: {response.status}"
     )
-
-
-def test_route_error_message_is_logged(mocker, test_client):
-    expected_error_msg = "400 Request missing 'barcode' parameter"
-
-    mocked_logger = mocker.spy(main.logger, "info")
-
-    response = test_client.get("/start_recording")
-    assert response.status == expected_error_msg
-
-    assert expected_error_msg in mocked_logger.call_args[0][0]
