+++ tests/test_process_monitor/test_process_monitor_handling_communication_from_server.py
@@ -1,10 +1,13 @@
 # -*- coding: utf-8 -*-
+import datetime
 import tempfile
 from uuid import UUID
 
+from freezegun import freeze_time
 from mantarray_desktop_app import BUFFERING_STATE
 from mantarray_desktop_app import CALIBRATING_STATE
 from mantarray_desktop_app import LIVE_VIEW_ACTIVE_STATE
+from mantarray_desktop_app import RECORDING_STATE
 from mantarray_desktop_app import START_MANAGED_ACQUISITION_COMMUNICATION
 import numpy as np
 from stdlib_utils import invoke_process_run_and_check_errors
@@ -291,3 +294,51 @@ def test_MantarrayProcessesMonitor__check_and_handle_server_to_main_queue__handl
         test_process_manager.get_values_to_share_to_server()["system_status"]
         == LIVE_VIEW_ACTIVE_STATE
     )
+
+
+@freeze_time(
+    datetime.datetime(
+        year=2020, month=11, day=16, hour=15, minute=14, second=44, microsecond=890122
+    )
+)
+def test_MantarrayProcessesMonitor__check_and_handle_server_to_main_queue__handles_start_recording__in_hardware_test_mode__by_passing_command_to_file_writer__and_setting_status_to_recording__and_updating_adc_offsets(
+    test_process_manager, test_monitor
+):
+    monitor_thread, _, _, _ = test_monitor
+
+    test_process_manager.create_processes()
+
+    server_to_main_queue = (
+        test_process_manager.queue_container().get_communication_queue_from_server_to_main()
+    )
+    expected_timepoint = 55432
+    adc_offsets = dict()
+    for well_idx in range(24):
+        adc_offsets[well_idx] = {
+            "construct": 0,
+            "ref": 0,
+        }
+
+    communication = {
+        "communication_type": "recording",
+        "command": "start_recording",
+        "is_hardware_test_recording": True,
+        "adc_offsets": adc_offsets,
+    }
+    put_object_into_queue_and_raise_error_if_eventually_still_empty(
+        communication, server_to_main_queue
+    )
+    invoke_process_run_and_check_errors(monitor_thread)
+    assert is_queue_eventually_empty(server_to_main_queue) is True
+    main_to_fw_queue = (
+        test_process_manager.queue_container().get_communication_queue_from_main_to_file_writer()
+    )
+    confirm_queue_is_eventually_of_size(main_to_fw_queue, 1)
+
+    actual = main_to_fw_queue.get(timeout=QUEUE_CHECK_TIMEOUT_SECONDS)
+    assert actual == communication
+
+    shared_values_dict = test_process_manager.get_values_to_share_to_server()
+    assert shared_values_dict["is_hardware_test_recording"] is True
+    assert shared_values_dict["system_status"] == RECORDING_STATE
+    assert shared_values_dict["adc_offsets"] == communication["adc_offsets"]
